A notebook that emulates the TIP
This notebook provides an example emulation (and assessment of
its quality) of the TIP RT model using the NCEO `gp_emulator`
package. Note that I haven't tested the derivatives, as the
built in step size for the finite difference approximation in the
`gp_emulator` codebase results in rubbish results.  They do work,
but I'd need to update the  emulators so that it can all be
shown with the same framework!