Single band emulators
The emulators are now produced for the vis and nir broadbands
using only the relevant parameters (e.g. NIR or VIS spectral
parameters). This results in a simpler emulation, but you still
need quite a few training samples to get the gradient right,
although this might be an artefact of the finite difference
comparison.