Tipping the TIP into eoldas_ng
This is the first attempt at putting the TIP in eoldas. Two main
things need to be added:

1. A TIP state operator (e.g. defining units, parameter boundaries,
transformations, etc.)
2. A modification of the time series GP observation operator to cope
with just two GPs for each of the two bands. This should be useful
for other types of problems, such as passive (or active) microwaves.

This is the first commit, still very incomplete.