A tool to download, store and query albedo
This script will download MODIS C5 albedo data from the SOAP server
using `suds`, store it in an sqlite3 database and query the database.
The script appears to work OK, and this committed version will loop
over the `fluxnet_sites.txt` file to download data for a number of
FLUXNET sites. I'm not sure yet of whether it's working or not,
but there's no reason to think it won't work.