Calculating the TIP cost using GPs
I have modified the `der_cost` method to calculate the cost
function using the same GP for the VIS and NIR bands. The code is
yet to be tested, but it improves on efficiency by doing all the
timesteps in one go.

Apart from testing, I need to modify the uncertainty calculations,
which are based on finite differences. This is clearly a pain in
the arse, but it might be achieved just by changing the code to
do what it was doing before and running a ton of GPs sequentially.